requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.0.7
djangorestframework==3.15.2
filelock==3.16.1
gdown==5.2.0
idna==3.10
numpy==2.1.3
pandas==2.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
six==1.16.0
soupsieve==2.6
sqlparse==0.5.1
tqdm==4.67.1
tzdata==2024.2
urllib3==2.2.3
